+++ src/control-plane/backend/lambda/api/test/api/application.test.ts
@@ -74,7 +74,7 @@ describe('Application test', () => {
                             Action: 'Create',
                             Parameters: [],
                             StackName: 'clickstream-ingestion1',
-                            TemplateURL: 'https://xxx.com',
+                            TemplateURL: 'https://example.com',
                           },
                         },
                         End: true,
@@ -152,7 +152,7 @@ describe('Application test', () => {
                             Action: 'Create',
                             Parameters: [],
                             StackName: 'clickstream-ingestion1',
-                            TemplateURL: 'https://xxx.com',
+                            TemplateURL: 'https://example.com',
                           },
                         },
                         End: true,
@@ -755,7 +755,7 @@ describe('Application test', () => {
                             Action: 'Create',
                             Parameters: [],
                             StackName: 'clickstream-ingestion1',
-                            TemplateURL: 'https://xxx.com',
+                            TemplateURL: 'https://example.com',
                           },
                         },
                         End: true,
