+++ src/control-plane/backend/lambda/api/test/api/pipeline.test.ts
@@ -2772,21 +2772,20 @@ describe('Pipeline test', () => {
     tokenMock(ddbMock, false);
     projectExistedMock(ddbMock, true);
     dictionaryMock(ddbMock);
-    createPipelineMock(ddbMock, kafkaMock, redshiftServerlessMock, redshiftMock,
-      ec2Mock, sfnMock, secretsManagerMock, quickSightMock, s3Mock, iamMock, {
-        publicAZContainPrivateAZ: true,
-        subnetsCross3AZ: true,
-        subnetsIsolated: true,
-        update: true,
-        updatePipeline: {
-          ...KINESIS_DATA_PROCESSING_NEW_REDSHIFT_UPDATE_PIPELINE_WITH_WORKFLOW,
-          tags: [
-            { key: 'customerKey1', value: 'tagValue1' },
-            { key: 'customerKey2', value: 'tagValue2' },
-            { key: BuiltInTagKeys.AWS_SOLUTION_VERSION, value: MOCK_SOLUTION_VERSION },
-          ],
-        },
-      });
+    createPipelineMock(mockClients, {
+      publicAZContainPrivateAZ: true,
+      subnetsCross3AZ: true,
+      subnetsIsolated: true,
+      update: true,
+      updatePipeline: {
+        ...KINESIS_DATA_PROCESSING_NEW_REDSHIFT_UPDATE_PIPELINE_WITH_WORKFLOW,
+        tags: [
+          { key: 'customerKey1', value: 'tagValue1' },
+          { key: 'customerKey2', value: 'tagValue2' },
+          { key: BuiltInTagKeys.AWS_SOLUTION_VERSION, value: MOCK_SOLUTION_VERSION },
+        ],
+      },
+    });
     cloudFormationMock.on(DescribeStacksCommand).resolves({
       Stacks: [
         {
