+++ test/control-plane/click-stream-api.test.ts
@@ -755,9 +755,13 @@ describe('Click Stream Api ALB deploy Construct Test', () => {
               'events:PutRule',
               'events:ListTargetsByRule',
               'events:PutTargets',
+              'events:TagResource',
+              'events:UntagResource',
               'sns:CreateTopic',
               'sns:Subscribe',
               'sns:SetTopicAttributes',
+              'sns:TagResource',
+              'sns:UntagResource',
             ],
             Effect: 'Allow',
             Resource: '*',
